requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
Faker==37.1.0
iniconfig==2.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.0
tzdata==2025.2
gunicorn